pyproject.toml
[project]
name = "assistantdefrancais"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.32.1",
    "opentelemetry-distro>=0.53b1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.32.1",
    "opentelemetry-exporter-otlp-proto-http>=1.32.1",
    "opentelemetry-exporter-prometheus>=0.53b1",
    "opentelemetry-instrumentation-aiohttp-client==0.53b1",
    "opentelemetry-instrumentation-aiohttp-server==0.53b1",
    "opentelemetry-instrumentation-asyncio==0.53b1",
    "opentelemetry-instrumentation-dbapi==0.53b1",
    "opentelemetry-instrumentation-httpx==0.53b1",
    "opentelemetry-instrumentation-logging==0.53b1",
    "opentelemetry-instrumentation-sqlite3==0.53b1",
    "opentelemetry-instrumentation-threading==0.53b1",
    "opentelemetry-instrumentation-urllib==0.53b1",
    "opentelemetry-instrumentation-wsgi==0.53b1",
    "opentelemetry-sdk>=1.32.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
]


[tool.pytest.ini_options]
testpaths="tests"
asyncio_default_fixture_loop_scope="session"
filterwarnings = [
    "ignore:'audioop' is deprecated"
]
